workertest: rename :Echo => :EchoJava.
the java worker action just started failing in a clean repo. the error
message says that bazel couldn't find the `java` binary in the echo
runfiles.

my hunch is that this is a horrible case-insensitive filesystem bug.
there was a target named `echo` (of type echo()) until 302f3cf, and i
notice traces of it in my bazel-bin. this would probably be fixed by a
bazel clean, but just renaming the target is faster (and works).